Working draft of menu
