+++ cmd/help.go
@@ -1,65 +1,66 @@
 package cmd
 
-// import (
-// 	"image/color"
-// 	"strings"
+import (
+	"image/color"
+	"strings"
 
-// 	"github.com/hajimehoshi/ebiten/v2"
-// 	"github.com/hajimehoshi/ebiten/v2/text/v2"
-// )
+	"github.com/hajimehoshi/ebiten/v2"
+	"github.com/hajimehoshi/ebiten/v2/text/v2"
+)
 
-// // HelpSection represents a static section in the menu to display game controls.
-// type HelpSection struct {
-// 	name     string
-// 	controls []string
-// }
+// HelpSection represents a static section in the menu to display game controls.
+type HelpSection struct {
+	name     string
+	controls []string
+}
 
-// func (hs *HelpSection) Name() string {
-// 	return hs.name
-// }
+func (hs *HelpSection) Name() string {
+	return hs.name
+}
 
-// func (hs *HelpSection) Value() interface{} {
-// 	return nil // HelpSection has no modifiable value
-// }
+func (hs *HelpSection) Value() interface{} {
+	return nil // HelpSection has no modifiable value
+}
 
-// func (hs *HelpSection) SetValue(val interface{}) error {
-// 	return nil // HelpSection cannot be modified
-// }
-// func (hs *HelpSection) Render(screen *ebiten.Image, x, y int, selected bool) {
-// 	// Colors for the bindings and descriptions
-// 	bindingColor := color.RGBA{255, 255, 0, 255}       // Yellow for the keybinding
-// 	descriptionColor := color.RGBA{255, 255, 255, 255} // White for the description
+func (hs *HelpSection) SetValue(val interface{}) error {
+	return nil // HelpSection cannot be modified
+}
 
-// 	// Render the help section title
-// 	title := "Help - Game Controls"
-// 	op := &text.DrawOptions{}
-// 	op.GeoM.Translate(float64(x), float64(y))
-// 	op.ColorScale.ScaleWithColor(color.White)
-// 	text.Draw(screen, title, loadedFont, op)
+func (hs *HelpSection) Render(screen *ebiten.Image, x, y int, selected bool) {
+	// Colors for the bindings and descriptions
+	bindingColor := color.RGBA{255, 255, 0, 255}       // Yellow for the keybinding
+	descriptionColor := color.RGBA{255, 255, 255, 255} // White for the description
 
-// 	// Adjust the starting y-coordinate for the controls
-// 	y += 40 // Adjust spacing as needed
+	// Render the help section title
+	title := "Help - Game Controls"
+	op := &text.DrawOptions{}
+	op.GeoM.Translate(float64(x), float64(y))
+	op.ColorScale.ScaleWithColor(color.White)
+	text.Draw(screen, title, loadedFont, op)
 
-// 	// Render each control in the list
-// 	for i, control := range hs.controls {
-// 		// Split the control into binding and description
-// 		parts := strings.SplitN(control, " - ", 2)
-// 		if len(parts) != 2 {
-// 			continue // Skip malformed entries
-// 		}
-// 		binding := parts[0]
-// 		description := parts[1]
+	// Adjust the starting y-coordinate for the controls
+	y += 40 // Adjust spacing as needed
 
-// 		// Render the keybinding
-// 		op = &text.DrawOptions{}
-// 		op.GeoM.Translate(float64(x), float64(y+(i*30))) // Adjust line height as needed
-// 		op.ColorScale.ScaleWithColor(bindingColor)
-// 		text.Draw(screen, binding, loadedFont, op)
+	// Render each control in the list
+	for i, control := range hs.controls {
+		// Split the control into binding and description
+		parts := strings.SplitN(control, " - ", 2)
+		if len(parts) != 2 {
+			continue // Skip malformed entries
+		}
+		binding := parts[0]
+		description := parts[1]
 
-// 		// Render the description in the second column
-// 		op = &text.DrawOptions{}
-// 		op.GeoM.Translate(float64(x+200), float64(y+(i*30))) // Adjust x+200 for column spacing
-// 		op.ColorScale.ScaleWithColor(descriptionColor)
-// 		text.Draw(screen, description, loadedFont, op)
-// 	}
-// }
+		// Render the keybinding
+		op = &text.DrawOptions{}
+		op.GeoM.Translate(float64(x), float64(y+(i*30))) // Adjust line height as needed
+		op.ColorScale.ScaleWithColor(bindingColor)
+		text.Draw(screen, binding, loadedFont, op)
+
+		// Render the description in the second column
+		op = &text.DrawOptions{}
+		op.GeoM.Translate(float64(x+200), float64(y+(i*30))) // Adjust x+200 for column spacing
+		op.ColorScale.ScaleWithColor(descriptionColor)
+		text.Draw(screen, description, loadedFont, op)
+	}
+}
